requirements.txt
automata-lib==6.0.2
click==7.1.2
Flask==1.1.4
Flask-Sockets==0.2.1
gevent==22.10.1
gevent-websocket==0.10.1
greenlet==1.1.3.post0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.1
networkx==2.8.7
pydot==1.4.2
pyparsing==3.0.9
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.5.0
bitarray==2.7.0
numpy
